update
- Criação da classe model Question e ItemChecklist ;
- Criação da classe model IdItemChecklist que será uma classe que unirá duas chaves estrangeiras como uma chave primaria ;
- Criação e implementação da classe e interface Service,  da interface DAO e da classe Controller do ItemChecklist;
- Alteração no metodo delete do profile;

+++ backend/src/main/java/br/com/nutrinotes/service/business/BusinessImpl.java
@@ -65,8 +65,13 @@ public class BusinessImpl implements IBusinessService {
 
 	@Override
 	public boolean delete(Integer id) {
-		dao.deleteById(id);
-		return true;
+		Optional<Business> b = dao.findById(id);
+		if (b.isPresent()) {
+			dao.deleteById(id);
+			return true;
+		}
+		System.err.println("Ocorreu um erro ao excluir a empresa");
+		return false;
 	}
